fix: move affirmation mock out of tests
