+++ features/self-care/components/affirmation/__tests__/AffirmationStoryModal.test.tsx
@@ -42,7 +42,7 @@ jest.mock("expo-linear-gradient", () => {
   };
 });
 
-jest.mock("react-native", () => require("./mockReactNative"));
+jest.mock("react-native", () => require("../mockReactNative"));
 
 jest.mock("react-native-safe-area-context", () => ({
   useSafeAreaInsets: () => ({ top: 0, bottom: 0, left: 0, right: 0 }),
