+++ features/self-care/components/affirmation/mockReactNative.js
